feat: support arbitrary fields when converting pointclouds
The extra fields are included if they are in the lists given by the
caller. Their datatype is kept, or a suitable datatype is used. By
suitable, we meant a datatype that can contain the corresponding data
types for this field in all the input files.

BREAKING CHANGE:

We now deal with extra fields in a generic way, so that users  can
include the fields they want, not only classification and intensity.
Additionnally, we don't include intensity and classification by default
any more.

Here's how to migrate

```diff
- convert(file, outfolder=folder)
+ convert(file, outfolder=folder, extra_fields=["intensity", "classification")
```
For cli invocations, please use multiple `--extra-fields` arguments.