feat(merger): rework it completely
Previously, we distinguished the pnts case and the other cases. In the
pnts case, the merger was subdividing again the root_tiles of the input
tilesets following a quadtree subdivision.

But a quadtree is but one possible relevant subdivision. Also, we could
just assume that the input tilesets are already correctly subdivided.
With this in mind, we *just* need to take a subset of points in the
input tilesets and be done with it. It is simpler and will work better
in the general case.

BREAKING_CHANGES: the parameter `force_universal_merger` of `merger.merge_from_files` has been removed. The merger automatically select the best strategy now.