feat: add a simple viewer to preview generated tilesets
The viewer.py file starts a local http server pointing on the provided
tilesets' parent directory, then it starts a browser pointing on the
simple 3dtiles viewer of the giro3d.org website.

Further work could improve that by not depending on an online service,
and supports piero, which would allow to display all these tilesets in
the same 3D scene.